Moved models.py in to the correct app folder
